+++ test/lib/actions/ActionRefinerTest.scala
@@ -1,13 +1,13 @@
 package lib.actions
 
 import com.gu.identity.play.{AccessCredentials, AuthenticatedIdUser, IdMinimalUser}
-import lib.TestUsers
 import org.scalatest.mockito.MockitoSugar._
 import org.scalatest.{MustMatchers, WordSpec}
 import play.api.mvc.Results._
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import play.api.http.Status
+import services.TestUserService
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
@@ -21,7 +21,7 @@ class ActionRefinerTest extends WordSpec with MustMatchers {
   "PrivateAction" should {
 
     "include Cache-control: private" in {
-      val actionRefiner = new ActionRefiners(_ => None, "", "", mock[TestUsers], stubControllerComponents())
+      val actionRefiner = new ActionRefiners(_ => None, "", "", mock[TestUserService], stubControllerComponents())
       val result = actionRefiner.PrivateAction(Ok("")).apply(FakeRequest())
       header("Cache-Control", result) mustBe Some("private")
     }
@@ -31,7 +31,7 @@ class ActionRefinerTest extends WordSpec with MustMatchers {
   "AuthenticatedAction" should {
 
     "respond to request if provider authenticates user" in {
-      val actionRefiner = new ActionRefiners(_ => Some(mock[AuthenticatedIdUser]), "", "", mock[TestUsers], stubControllerComponents())
+      val actionRefiner = new ActionRefiners(_ => Some(mock[AuthenticatedIdUser]), "", "", mock[TestUserService], stubControllerComponents())
       val result = actionRefiner.AuthenticatedAction(Ok("authentication-test")).apply(fakeRequest)
       status(result) mustEqual Status.OK
       contentAsString(result) mustEqual "authentication-test"
@@ -39,7 +39,7 @@ class ActionRefinerTest extends WordSpec with MustMatchers {
 
     "redirect to identity if provider does not authenticate" in {
       val path = "/test-path"
-      val actionRefiner = new ActionRefiners(_ => None, idApiUrl, supportUrl, mock[TestUsers], stubControllerComponents())
+      val actionRefiner = new ActionRefiners(_ => None, idApiUrl, supportUrl, mock[TestUserService], stubControllerComponents())
       val result = actionRefiner.AuthenticatedAction(Ok("authentication-test")).apply(fakeRequest)
 
       status(result) mustEqual Status.SEE_OTHER
@@ -53,13 +53,13 @@ class ActionRefinerTest extends WordSpec with MustMatchers {
     }
 
     "return a private cache header if user is authenticated" in {
-      val actionRefiner = new ActionRefiners(_ => Some(mock[AuthenticatedIdUser]), "", "", mock[TestUsers], stubControllerComponents())
+      val actionRefiner = new ActionRefiners(_ => Some(mock[AuthenticatedIdUser]), "", "", mock[TestUserService], stubControllerComponents())
       val result = actionRefiner.AuthenticatedAction(Ok("authentication-test")).apply(fakeRequest)
       header("Cache-Control", result) mustBe Some("private")
     }
 
     "return a private cache header if user is not authenticated" in {
-      val actionRefiner = new ActionRefiners(_ => None, idApiUrl, supportUrl, mock[TestUsers], stubControllerComponents())
+      val actionRefiner = new ActionRefiners(_ => None, idApiUrl, supportUrl, mock[TestUserService], stubControllerComponents())
       val result = actionRefiner.AuthenticatedAction(Ok("authentication-test")).apply(fakeRequest)
       header("Cache-Control", result) mustBe Some("private")
     }
@@ -68,7 +68,7 @@ class ActionRefinerTest extends WordSpec with MustMatchers {
 
   "AuthenticatedTestUserAction" should {
 
-    val testUsers = new TestUsers("test") {
+    val testUsers = new TestUserService("test") {
       override def isTestUser(displayName: Option[String]): Boolean = displayName.exists(_.startsWith("test"))
     }
